HFP: A2dp Suspended flags needs to reset on Audio server restart.
When audio server restarts, audio params need to be restored
and while setting the bluetoothSco to off and on, A2dpSuspended
and LeaSuspended flags are setting to true.
But need to set them to false before calling the sco off and
then set to true when setting the Sco to true.

CRs-Fixed: 3708954
Change-Id: Iff4412fb7bae2d830a99394d497ef597d58a1502